(ref) - MVC pattern of app.
