+++ src/bot.py
@@ -1,3 +1,5 @@
+import logging as log
+import os
 from dataclasses import dataclass
 
 from aiogram import Bot
@@ -6,35 +8,42 @@ from aiogram.types import ReplyKeyboardMarkup, KeyboardButton
 from environs import Env
 import utils
 
+from models.section import get_sections
 
 @dataclass
 class BotConfig:
+  _locale: str = 'ru'
+  supportedLocales = ("ru",)
+  sections = {}
+
   @property
   def locale(self):
     return self._locale
 
   @locale.setter
-  def locale(self, newVal):
-    if newVal not in self.supportedLocales:
-        raise f"{newVal} locale is not supported."
-    self._locale = newVal
+  def locale(self, val):
+    if val not in self.supportedLocales:
+      raise ValueError(f"{val} locale is not supported.")
+    self._locale = val
+    self.resources = utils.get_locale_res(val)
 
-  # Methods
+  def initialize_sections(self, directory: str):
+    self.sections = {section.name: section for section in get_sections(directory)}
+    log.info(self.sections)
+
   def __init__(self) -> None:
-    # TODO - find better way to define props
-    # Data
-    self.supportedLocales = ("ru")
-    # self.resources: dict[str, Any]
-
-    # NOTE - Default locale is "ru"
-    # DEFECT - Get first locale from supported, no hard-code
-    self.locale = "ru"
     self.resources = utils.get_locale_res(self.locale)
+    self.sections = {}
 
-# Global obj
+
+# Global object
 bot_config: BotConfig = BotConfig()
 
-# TODO - ugly
+src_dir = os.path.dirname(os.path.abspath(__file__))
+content_dir = os.path.join(src_dir, '../content')
+bot_config.initialize_sections(content_dir)
+
+
 env = Env()
 env.read_env()
